Fix Rubocop setup
While testing inheriting from the gem source I noticed strange behavior
in regards to how `Exclude` was behaving. After re-reading the docs it
turns out that the configuration file names are important:

> In `.rubocop.yml` and any other configuration file beginning with
> `.rubocop`, files and directories are specified relative to the
> directory where the configuration file is. In configuration files that
> don't begin with `.rubocop`, e.g. `our_company_defaults.yml`, paths
> are relative to the directory where `rubocop` is run.
>
> -- https://rubocop.readthedocs.io/en/latest/configuration/#includingexcluding-files

This means that even with the `inherit_mode` set locally for a project
all of the "standard" excludes were essentially being ignored; as they
were being taken relative to the gem's directory. Since we want this to
be relative for the project, as they are common configurations, we need
to use file names which do not start with `.rubocop`.

This moves the existing common Rubocop configurations to files using a
`common_` name prefix. Local testing confirms this resolves the file
exclusion issue. But, the project's local `.rubocop.yml` still needs to
include the `inherit_mode` directive as well for it to work properly.