Setup initial common RSpec configuration
This documents how we expect this gem to be used. This gem is meant to
centralize the standard RSpec configuration for our projects. In this
respect, including it is intended to essentially be a replacement for
`spec/spec_helper.rb`.

However, every project always has some sort of custom configuration
changes necessary that are specific to them (i.e.  project specific
helpers, custom `ENV` settings etc.). With that in mind we don't expect
projects to remove `spec/spec_helper.rb`. Instead we state that it
should require this gem as the first line and that it should mostly be
empty. Any custom project configuration should be included as normal
after requiring this gem.

When we setup this gem we setup the `spec/spec_helper.rb` based on the
core common configuration across our projects; or at least the desired
configuration. This moves that into the gem itself and then we follow
our own advise in regards to our local RSpec setup.

On the Rails side things are a little trickier. We want this gem to
focus more on the spec side and not necessarily the boiler plate Rails
environment setup. As the Rails specific setup is more likely to change
across releases we don't want to depend on it here in the gem. Thus our
instructions for the suggested `spec/rails_helper.rb` closely mirrors
the standard file; the only real change is instead of requiring
`rspec/rails` projects require `radius/spec/rails`. This is of course
after already requiring `spec/spec_helper.rb` which should require this
gem.

We still need to be concerned about the case where someone doesn't
follow the advise about loading `spec/spec_helper.rb`, which should
require this gem. We guard this by requiring this gem is when the Rails
config file is loaded.