Add Rails RSpec config
This is the default `RSpec.configure` setup for `spec/rails_helper.rb`.
While we don't always use fixtures the fixture configuration is included
for two reasons:

  - a few projects uses fixtures in a few places for speed / mass setup
  - the `use_transactional_fixtures` is necessary for the general DB
    cleanup transaction strategy

One thing that people tend to forget initially, but should be included
in virtually every Rails project, is cleaning up the cache between
specs. The default cache is usually either the null cache, which doesn't
require cleanup b/c it doesn't store anything, or an in-memory cache.
Both of those stores respond to `clear`. The other common test store is
the `file_store` which also supports clear. We tend to use Redis for our
cache and the Redis store is another one supporting `clear`. This is why
we include this hook in the setup.

This configures RSpec to exclude the new system tests by default. Per
the code comment this is to more closely mirror the default Rails setup
and to handle projects which do not install all dependencies by default
(thus breaking when system specs run).

We use a string for the exclusion filter to ease including the specs on
the command line. The RSpec configuration is smart enough to exclude
specs which use the symbol value `:system` when we used the string
`"system"` in the config. We can include them on the command line with:

    bin/rspec -t type:system

If we used the symbol value they would be excluded still:

  ```console
  $ bin/rspec -t type:system
  Run options:
    include {:type=>"system"}
    exclude {:type=>:system}

  Randomized with seed 17536

  Finished in 0.00081 seconds (files took 3.29 seconds to load)
  0 examples, 0 failures
  ```

It is possible to send a symbol on the command line but it requires
double colons:

    bin/rspec -t type::system

Similarly if set the exclude filter using a symbol, but use the string
variant on the command line everything is excluded as before:

  ```console
  $ bin/rspec -t type::system
  Run options:
    include {:type=>:system}
    exclude {:type=>"system"}

  Randomized with seed 2938

  Finished in 0.00082 seconds (files took 3.15 seconds to load)
  0 examples, 0 failures
  ```

Just based on our usage it seems people more often use the non-symbol
variant on the command line.