Document desired model factory behavior
In general, when it comes to factories our policy is to prefer explicit
over implicit. Specs should be written in a way that makes it clear what
is relevant, and not relevant, to the object/behavior under test.

Our goal is to create a "simple" model factory. We define "simple" as:

  - minimal feature set
  - easy to use public API
  - minimal meta-programming

An additional goal is that the factory should **not** be Rails specific.
However, given the number of factory gems and the popularity of a few we
make some concessions as far as naming goes to follow the common
patterns.

Specifically, we plan on supporting `build` and `create` methods for
instantiating objects. Ideally we'd not support `create` as it can lead
to odd issues with RSpec's `let` construct, which lazy evaluates the
associated block, and instances not being persisted when expected.
Instead we'd suggest people get in the habit of being explicit about
what needs to be "preloaded" using a `before` block and explicitly
calling `save!`:

  ```ruby
  let(:an_object) { build(Object) }

  before do
    an_object.save!
  end
  ```

But there are enough times when, due to save order dependencies, you
need to create something to provide to something else. And using a
`before` hook like this isn't a viable option. And calling
`tap(&:save!)` litters the code.

However, there are features common to other factory gems that we
intentionally want to omit. Again, this is to keep the factory simple,
but also to direct people away from a lot of common pitfalls :snake:
that those easy lead to.

  - Calling factories from within factories

    This has historically been a path leading people into a labyrinth of
    confusing object dependencies, slow test suites, and hidden state.
    All of this detracts from ease of development and spec
    comprehension.

    We choose to require people to be explicit about object relations
    and dependencies. Exposing state and being explicit in specs as to
    the data is required for functionality.

  - Named factories

    Our interpretation of this feature from other gems is that it's:

      - an attempt to mimic fixtures
      - provide codification of agile "personas"

    We omit it for the simple reason that is places greater distance
    between object state and the specs. Unless you are very familiar
    with all the intricacies of the named entity in question it can
    often be confusing to understand why a spec fails or why some
    additional state you setup doesn't match your expectation.

  - Inheriting/extending from another factory

    In our experience, this has limited benefit in the grand scheme of
    factories. If you look at other factory gems this is generally used
    in conjunction with named factories or is used in the name of
    [DRY](https://en.wikipedia.org/wiki/Don%27t_repeat_yourself) to
    group together repeated attributes.

    Our experience is that this tends to makes factory comprehension
    more difficult only to save a few lines of code.