Implement initial factory registration
Per the README, this sets up the `ModelFactory` module to support the
following public APIs:

  ```ruby
  Radius::Spec::ModelFactory.define_factory(
    "AnyClass",
    attr1: :any_value,
    attr2: :another_value,
  )

  Radius::Spec::ModelFactory.catalog do |c|
    c.factory "AnyClass", attr1: :any_value, attr2: :another_value
  end
  ```

I went back and forth a lot on what to call the block form of the
method. I settled on `catalog` for two main reasons:

  - I didn't want to use `configure` as I have a feeling we'll want to
    reserve that for future enhancements where we may actually configure
    different features / parts of the factory. Overloading `configure`
    with the interface for defining factories would potentially conflate
    these two concepts.

  - In the past we talked about creating a standalone factory gem. The
    more I looked at what we did there and what we may do here I have a
    feeling that's still a path we may take. It's likely that the
    external gem will use this `catalog` API and I wanted to reduce
    potential changes at that point. Though this is a guess so maybe
    it'll be wrong :grimacing:.

This implementation is the bare bones basic implementation we've been
using in our apps; with the exception of how we store the class names.
The change in class names is to allow a faster factory definition / load
time. As there is generally a single file defining all of the factories,
for large projects (i.e. Rails apps), using the actual class would
essentially cause the entire project to be loaded. This is troublesome
for isolated testing where only one or two factories are required (which
may not even depend on Rails; remember this is `ActiveRecord` agnostic).

For simplicity in getting the initial version released we are using a
simple hash to store the templates. This will likely change in the
future. However, we need to be able to clear the templates internally
for testing. To avoid having to call `send` we make the API publicly
accessible, but mark it as a "private" API. This lets clients know that
if they decide to tie themselves to this method or the implementation it
is at their own risk.

We duplicate, and freeze, the provided attribute hash to preserve the
registered template definition. However, we've chosen not to duplicate
the actual registered attribute values. We strongly considered
duplicating unfrozen values but ultimately decided against it for a few
reasons:

  - in all of our apps that use the existing factory implementation
    we've not had a problem which would be solved by duplicating the
    values

  - `dup` is shallow by default or values which are hashes or otherwise
    have state can still be mutated

    We can't freeze these objects because there's no guarantee there
    won't be internal object state that is lazy evaluated (`Time` had
    this problem in the past). Also, it's likely non-proc generators may
    have internal state cannot be frozen.

  - there is a rare edge case where "constants" are defined but
    intentionally mutable

    This occurs in a few places in Rails. The constants are more used as
    global state storage allowing people to add / remove things from the
    sets as configuration. While we believe constants should not be
    mutable we have to prepare for cases like this.

    ```ruby
    SOME_CONFIG = %w[ default values ]

    Radius::Spec::ModelFactory.define_factory(
      "SomeClass",
      arg: SOME_CONFIG,
    )

    # ... somewhere else in the app (maybe an initializer)
    # it's likely expected the "SomeClass" factory now includes "foobar"
    SOME_CONFIG << "foobar"
    ```