Implement model factory instance creation helpers
This implements the common `build` and `create` helpers for factories.
The key features here are:

  - "safe" transformation of registered default attributes

    This duplicates the values to avoid array and hash modifications
    leaking across instances. This only shallow duplicates the values as
    that's generally enough. There are some edge cases with
    arrays of arrays, hashes with mutable values, and generally mutable
    objects. However, there is no standard deep dup implementation (but
    Rails does have one so maybe we'll add it in the future).

    We do not duplicate values which are frozen. This is a minor
    optimization to help cut down on extra object creations. In general,
    when an object is frozen the expectation is it cannot be modified;
    again there are exceptions with nested objects but in general
    there's an expectation that the frozen object has already taken
    responsibility to freeze it's mutable state or protect the mutable
    state when it was frozen.

  - registered default attribute values which are procs get `call`'d

    Because there are lots of objects which can respond to `call`, many
    of which are not generators, we need to be careful what we call
    `call` on. When we wrote the README we specified we were only going
    to support `Proc`. Other generators can be use by either converting
    them to procs (with `to_proc`) explicitly or wrapping them in a proc
    (e.g. `-> { some_generator.call }`).

    Often generators are slow and some are expensive (such as hitting
    the database - i.e. someone included another factory in it). Due to
    this we don't want to just blindly run every generator every time.
    Instead we only run the generators for those registered attributes
    which are not replaced by custom attributes provided to the factory
    helper.

  - remove any optional registered attributes which were not replaced by
    custom attributes provided to the factory helper

A common Ruby idiom, and one that `ActiveRecord` implements, is to yield
the instantiated object on the call to `new`. Sometimes this is simply
called at the end of `initialize` other times it can be called before a
finalization step (such as freezing) to allow final modifications to an
object. We support this idiom by passing any block provided to the
factory helpers through to `new`. Since any method can talk a block in
Ruby, even if it doesn't use it, this is a benign action to take even
for classes which don't support this idiom.

At this time if an underlying class doesn't use the block it is just
dropped on the floor. It would be nice to have the factory helpers
"standardize" the idiom by ensuring the block is always called. However,
this may not be straight forward and is not a feature we have supported
in the past. We've left a TODO note in the code to reconsider this again
in the future.

The `create` implementation is essentially just `build` with a
persistence call. This supports the standard Rails persistence method
`save!`. The "bang method" is used over the regular `save` so that
things fail fast and loudly; the `save!` version raises an exception on
model validation failure. At this point there isn't another persistence
API we need to support, but it wouldn't be too difficult to extend this
to support another.