Hook into RSpec to load model factory
This adds hooks for RSpec to auto load the model factory via metadata.
Based on the usage in our current projects we don't _always_ need a
model factory for each of these spec types. However, we use it often
enough that it's worth including by default. With the new change to
support lazy evaluation of the actual project/app class code loading the
factory definitions should be fast enough that always including it in
these cases should be acceptable.

As each of these `when_first_matching_example_defined` blocks is
identical we could have setup an array to iterate over all of the
metadata permutations. However, we fully plan to add other features in
the future which will only apply to a subset of these. Having each
metadata in a separate block makes it easier to see everything that is
being included for a particular metadata without having to parse through
a bunch of separate iteration blocks.