Auto load a project's spec/support/model_factories
Our current idiom for where to store model factories is in the
`/spec/support/model_factories.rb` file. We use a single file to make it
easy to find all of the factories to load. It also means there's no
additional boiler plate necessary to pick a directory and then load /
require all files within it.

When RSpec is run it adds the local project's `/spec/support` directory
to the load path. This means a require from anywhere with the path
`support/model_factories` _should_ load from the path RSpec includes. As
the last bit of magic (ﾉ◕ヮ◕)ﾉ*:･ﾟ✧  this attempts to require that file
by default when the factory code is loaded.

We do this when the factory code is loaded, and not in the RSpec hooks,
to handle those projects which want to require the model factory feature
directly. This way those projects can setup a `support/model_factories`
in their load path and get the factories loaded when the feature loads.