Fix error with Rails RSpec configuration
This fixes the following error:

    NameError:
      undefined local variable or method `config' for main:Object

Not sure how this slipped past the initial round of external testing.
Since we don't have a setup for integration/acceptance testing the Rails
specific code parts yet it's not surprising something like this slipped
through. Though it is proof that we should include a way to perform
automated Rails integration testing.