Fix model factory building custom string key attrs
The existing model factory supports registering templates with either
string or symbol keys. The factory converts all keys to symbols for
consistency. This means when a model instance is built using the factory
the expectation is that symbol keys will be provided. Further, when
building a model with custom attributes the expectation is that the
custom attributes overwrite the registered template prior to
the model initialization.

Issue
-----

The following model class fails to see custom attributes defined using
string keys:

  ```ruby
  class Foo
    def initialize(attrs = {})
      @something = attrs.fetch(:something)
    end

    # ...
  end
  ```

Cause
-----

The :bug: is caused due to the builder not converting the custom
attribute keys into symbols. Thus the model is sent a hash which
contains a combination of string and symbol keys:

  ```ruby
  {
    :something    => "any_value",
    :another_attr => "another_value",
    "something"   => "custom_value",
  }
  ```

While we did not have a test specifically covering the message passing
contract, the existing test _"merges the custom attributes with the
registered ones"_ will pass even if we changed the custom attributes to
use string keys. This is because Ruby, in an effort to support backwards
compatibility for older code upgrading to kwargs, is "kind" enough to
treat the string key as a kwarg. And since Ruby hashes are ordered, and
the string key appears last, it is used as the kwarg.

Why support custom attributes with string keys?
-----------------------------------------------

The main reason is convenience in specs.

Depending on the app/spec type, I'm looking at you Rails, sometimes
custom data comes from a JSON source. When Ruby parses JSON into a data
hash it defaults to string keys. While JSON parsing supports converting
to symbol keys adding the explicit flag is less common (at least among
our apps).