Extract method refactor
This is mainly to make Rubocop happy. However, it nicely groups all of
the logic involved specifically with attribute merging into one place.

Note
----

This extraction helps define a further separation between what "build"
is responsible for and what something else should be responsible for.
Seeing the growing number of private API helper methods it seems like we
are missing an abstraction or two. We should explore this further soon.