Fix failing Travis CI setup
The builds are failing due to a strange error where rake cannot find
other dependencies.

    bundler: failed to load command: rake
    LoadError: cannot load such file -- bundler/dep_proxy

It appears this is a known issue (travis-ci 9062) which links to a Ruby
Gem issue (rubygems 2055).

This implements a hopefully simple fix which ensures bundler is updated
on the system, and that the installed version of Rails has the bundler
gem available.