Fix warning about overriding parameter
This fixes the following warning:

  radius-spec/common_rubocop_rails.yml: Metrics/LineLength:IgnoredPatterns
  overrides the same parameter in common_rubocop.yml

Even updating the local `.rubocop.yml` file to use
`common_rubocop_rails.yml` instead of `common_rubocop.yml` didn't
produce the issue. Only when run from another project was this
discovered :weary:

I wonder if the failure to detect this has to do with the local repo
containing the files directly while other projects include them via the
gem. It would be good in the future to have some way to test the Rails
setting integrations like this.