Lock Rubocop to a minor release version in gemspec
We've already encountered one fairly major issue with the `Include`
setting on all cops related to updating Rubocop. Additionally, as we
look to add more configurations, which may include new cops in the
future, we want to ensure projects have compatibility. In the past we've
tended to focus on bumping this gem in our projects and not Rubocop.
This ensures we keep everything in sync.

The downside of this change is that now installing the gem installs
Rubocop.