Bump Rubocop to 0.54.0
This Rubocop version added two new cops: `Rails/HttpStatus` and
`Style/UnpackFirst`. For Rails we already prefer the named status so we
leave the Rubocop defaults. The new style cop for `unpack` is actually
also a performance cop based on our testing. The default is to use
`unpack1` which is also the most performant so we won't configure it
either.