Bump Rubocop to 0.55.0
This Rubocop version added two new cops: `Performance/UnneededSort` and
`Lint/SafeNavigationConsistency`. Both of these cops match our current
code style so we accept the default configuration for them.

Additionally, this Rubocop version adjusted how the cop
`Naming/HeredocDelimiterNaming` parses the delimiters. It fixed the
regexp so that it only focuses on full matches of EO*. To keep current
with that, we update our customization for allowing `EOF` to match.