Bump Rubocop to 0.56.0
This Rubocop version adds several new cops:

  - `Performance/InefficientHashSearch`
  - `Rails/RefuteMethods`
  - `Rails/AssertNot`
  - `Lint/ErbNewArguments`
  - `Lint/SplatKeywordArguments`

We do not use assert style testing so neither of the new Rails cops
affects us. The lint cops are meant as preparation from Ruby 2.6 so
we're accepting the defaults. Finally, the performance cop seems
straight forward so no configuration to do with it either.

The `Layout/ClosingParenthesisIndentation` cop was also adjusted which
fixes an issue we logged. This is why we remove it's disabling in our
common config. Finally, this version also changed how it looked up
files; which was the reason we decided to drop `Include` back in
dfa52bba.