Add service integration for review :dog: check
This runs Rubocop adding the output as a check for the PR. Now that it
runs as a check we are removing it from failing the build, except for
lint issues; those will fail the build still. In general we expect all
Rubocop violations to be fixed, or configured to be allowed, so having
the Rubocop check fail is a sufficient flag. We let lint still fail the
build because they generally catch things which are most likely errors
and should be corrected.