Disable `Style/RedundantReturn` cop
In Ruby every method returns a value. Implicitly this is the value of
the last line of the method. This means using `return` is often
redundant.  However, there isn't anything inherently wrong about doing
so. In fact, in some cases it can help with a consistent style in a
method:

  ```ruby
  def transform(value)
    return value.call if value.is_a?(Proc)
    return value if value.frozen?
    return value.dup
  end
  ```

Other times it can add context to a seamingly oddly place value:

  ```ruby
  def munge(data)
    data.slice! 5
    return nil
  end
  ```

We often omit the explicit `return` in our code. Though sometimes we
include it for improved contextual clues and we don't want Rubocop to
complain for those cases.