Add line break to long details info
When including the extra details Rubocop 0.58 appends it onto the
initial line listing the issue. It then preserves the line breaks in the
YAML file when displaying the remainder of the message. To keep our
longer details information formatted per the YAML files we add an
initial line break to separate it from the cop listing.