Fix `NoMethodError` on RSpec 3.8 for fixture paths
This fixes the following error when using the `fixture_file_upload`
helper in request specs:

    NoMethodError:
      undefined method `strip' for #<Pathname:0x00007fab0a970e70>

The issue is caused by a change in RSpec 3.8 to support relative paths
when no fixture path is set. This resolves the issue by converting the
path to a string when setting it on the RSpec config.