Add more functional block method names
The `default_scope` method is used by Active Record and the
`filter_sensitive_data` is used in VCR configs. While the return value
of the these blocks are not used at the method call site, they are used
as dynamic methods / procs. So the values are imported and used at a
later time.