Rename model helper `create` to `build!`
This is response to working with the newly enabled `Rails/SaveBang`
Rubocop setting. As with `Rails/SaveBang` using `build!` guarantees
`save!` is called. Unfortunately, `Rails/SaveBang` does not flag either
of the following:

    build(SomeClass, &:save)
    build(SomeClass).tap(&:save)

This has made me re-think our original position on `create` slightly
since it provides a similar guarantee. In Rails world, my main issues
with `create` are:

  - it tends to be over used causing unnecessary DB calls when `build` /
    `new` would work

  - it can causing surprising behavior with `let` when people are
    expecting the object to exist at a certain point:

    ```ruby
    let(:obj) { create(Obj) }

    it "will fail because obj does not exist" do
      expect {
        obj.destroy!
      }.to change {
        # When this is first called `obj` has not be called and thus
        # does not exist yet
        Obj.count
      }.by(-1)
    end
    ```

    This then leads to people either using `let!` or simply reference
    the named variable in a bare context:

    ```ruby
    before do
      obj
    end

    it do
      obj
      # ... rest of spec
    end
    ```

There isn't much we can do about the first issue. However, I'm hoping by
calling this `build!` people will prefer to type the shorter (yes it's
only one character) `build` instead ¯\_(ツ)_/¯

As for the second issue. I think education on using `build` and `save!`
in the `before` block is helpful. Then also in the future writing a
custom Rubocop check for `build!` being the value of a `let` will be
enough.

This leaves `create` in place for our legacy apps, but at this point
really should be switched out for `build!` so we can remove it.