Add `using_tempfile` file stub helper
This adds a file stub helper we've used in a few projects. It allows us
to co-locate file stub data with the specs. It also allows us to freely
change the file stub data between specs without worry that there's
another spec which may be affected (which is a concern with file
fixtures).