Customize Rubocop Layout/AlignHash
In preparation for updating to Rubocop 0.60+ we need to customize this
cop. Unfortunately, this release changed how the cop treated the value
alignment, without new options to customize that portion. There is an
open issue which we will monitor to see if we can retain the legacy
behavior. Additionally, in response to this open issue Rubocop will be
disabling the cop by default.

To continue to get some value out of alignment checks we've decided to
stick to the `key` style. Across repos this setting requires less
changes for us. Lastly, we force enable the style so that we're ready
for when the next release is made.