Test against Ruby 2.6
We haven't update the default Rubocop ruby version yet because 2.6 was
just released. We tend to wait a little bit for the bugs to get found
before updating to new versions. But we want to ensure any code changes
made are 2.6 compatible.

One of the new improvements in Ruby 2.6 is performance around calling
blocks in methods. This also includes an update to the benchmarks to
show the change in Ruby 2.6. As can be seen here the improvement for the
"pass through block" case is significant.