Update Travis CI config
By default Travis CI uses bundler 1.x. Bundler 2.x is available now and
our attempts to update the version forced the update to 2 but that broke
the build. This removes that force update to remain on the default for
Travis. This also removes the deprecated `sudo` key and uses the newly
named `jobs` key.