Update to Rubocop 0.71.0
Version 0.71.0 is the last version before the Rails cops were moved into
a new repo. This is also the version that Ruby 2.7 was first permitted
to be used as a target version.

We're pushing to update all of our apps to Ruby 2.7 so we're going to be
setting that as the default `TargetRubyVersion` moving forward.

This enables two additional lint checks which match our teams preferred
styles. Similarly, we disable two checks because they are either overly
opinionated for something we do not care about or do not provide us
enough configuration options to support our style.

Lastly, we slightly increase the complexity counts to give us a bit more
wiggle room in our code. Most of our code is for Rails apps which can
sometimes require the extra room. We considered moving these changes to
the Rails only configuration, but they were small enough increases we
decided to make them part of the common configuration.

A note regarding the change from Layout/FirstParameterIndentation to
Layout/IndentFirstArgument. The current version of Rubocop claims:

  Error: The `Layout/FirstParameterIndentation` cop has been renamed to
  `Layout/IndentFirstArgument`.

However, based on the most recent version of Rubocop cop, the original
check still exists. So at some point between where we are now and the
most recent version this check split. One of the checks which came out
of it was named similar to the name we had been using. When we reach
that point we'll circle back around and re-evaluate the checks.

Co-authored-by: Aaron H <armahillo@gmail.com>
Co-authored-by: Benjamin Reynolds <breyno127@gmail.com>
Co-authored-by: Chris Hoffman <choffman@radiusnetworks.com>
Co-authored-by: James Nebeker <jnebeker@radiusnetworks.com>