Bump to rubocop 0.77.0
Upgrade rubocop from version 0.75.1 to version 0.77.0. The biggest
breaking changes came from version 0.77.0 where many cops were re-named.
I started by trying to make sense of the issue linked to in the
CHANGELOG:

https://github.com/rubocop/rubocop/blob/master/CHANGELOG.md#0770-2019-11-27
https://github.com/rubocop/rubocop/issues/7077

But in the end I realized it was faster to just upgrade and let rubocop
cop warn me of any unknown cops that needed to be re-named in version
0.77.0. I kept running this command in my local `radius-spec` repo until
the errors stopped:

```
$ bin/rubocop --list --config .rubocop.yml
```