Switch to disabling Lint/SuppressedException
It isn't called out in the CHANGELOG but it looks like the
`Lint/HandleExceptions` cop was re-named to `Lint/SuppressedException`.
Disabling the new cop name causes the Rubocop check to pass again.

Note: In all of these cases we have a comment explaining why we're
suppressing the exception, when we upgrade to rubocop  0.81.0 the
`AllowComments` for this cop will be true by default so we should be
able to remove these disablings of the cop.