Update to Rubocop 0.82.0
Version 0.80.1 noted the following change in the notes:

    7765: When warning about a pending cop, display the version with the
    cop added.

That change would require us to explicitly opt-in, or out, of new checks until the
next major version. When the major version flips over, we'd be able to drop
those explicit opt-in/out entires if they matched the default status.

However, in 0.82.0, an additional change was made that allows us to
revert to the old behavior of opting into the default behavior for new
checks:

    7850: Make it possible to enable/disable pending cops.

We're choosing to take advantage of this new setting so we do not have
to retain a large "default opt-in" list of checks. We are already going
version by version so that we can understand each new check and any new
configuration settings. So the explicit opt-in/out is actually more work
for us during this process.

Co-authored-by: Benjamin Reynolds <breyno127@gmail.com>