Update to Rubocop 0.90.0 and Rubocop-Rails 2.9.1
We crossed referenced our code bases and the vast majority of class
methods are all defined by `def self.method_name`. To remain consistent
with that style we're enabling the check for it.

In a turn of irony, this repo actually uses `class << self` style for
the main declaration of class methods. We used it in this repo because
we needed to call `alias_method` within the singleton class context. We
prefer `alias_method` to be called immediately after the canonical
method definition. We didn't like the look of having `class << self`
used for a single `alias_method` in the middle of multiple class method
definitions and we didn't want to locate the lone `alias_method` away
from the canonical definition just for a style check.

However, once we drop support for Ruby < 2.7 we can switch back to the
`def self.method_name` style. When we do we can use
`singleton_class.alias_method` to avoid usage of `class << self`.

Co-authored-by: Aaron Hill <armahillo@gmail.com>
Co-authored-by: JC Avena <jcavena@gmail.com>
Co-authored-by: Sam Kim <sam@radiusnetworks.com>