Exclude case from duplicate branch lint check
In this case we always want to prevent CI from recording cassettes. It
just so happens, that's also our current default behavior we're setting.
We don't want to lose the intent of CI's behavior by combining the
duplicate branches. Further, we want the `when` to act as an ordering
case check. As such the CI behavior should always trump the
configuration based on the number of specs running.

This is a case of coincidental duplication and not true duplication. As
such the duplicate branch lint is incorrect in flagging this; so we are
disabling it at the call site.